Add files via upload
In this project. I made a gesture/Virtual Mouse .In this i made a gesture mouse in which the cursor moves according to your hand. In this i have used open cv and color calibration in which i had used three color blue,yellow,red in which each color performe different action indivdual and combine and performe accordingly.
The action that can be performed are
1-Right click
2-left click
3-scroll
4-cursor movement